Now has default width
Added a variable to represent the default width for images. This
only works for images, because I can't figure out how to easily
resize animations. It now also stores the original image so that
later, when I write it, I can save the original image and not a
resized version.